Agregando ediciones, carrito accesible desde donde sea

+++ app/src/main/java/com/agroconecta/mobile/data/model/CartItem.kt
@@ -2,8 +2,9 @@ package com.agroconecta.mobile.data.model
 
 data class CartItem(
     val product: Product,
-    val quantity: Int
+    val quantity: Double
 ) {
+
     val totalPrice: Double
         get() = product.price * quantity
 }
