+++ app/src/main/java/com/agroconecta/mobile/data/remote/dto/PurchaseDto.kt
@@ -6,8 +6,8 @@ data class PurchaseDto(
     val productName: String,
     val farmerId: Int,
     val buyerId: Int,
-    val quantity: Float,
-    val price: Float,
+    val quantity: Double,
+    val price: Double,
     val totalPrice: Double,
     val status: String,
     val createdAt: String
